Refactoring of the project
