+++ src/main/java/fillstrategy/impl/FillingListAuto.java
@@ -1,0 +1,12 @@
+package fillstrategy.impl;
+
+import fillstrategy.FillingList;
+
+import java.util.List;
+
+//public class FillingListAuto implements FillingList {
+//    @Override
+//    public void fillList(List list) {
+//
+//    }
+//}
