+++ src/main/java/fillstrategy/impl/FillingListFromFile.java
@@ -1,4 +1,4 @@
-package strategy.impl;
+package fillstrategy.impl;
 
 public class FillingListFromFile {
 }
