+++ src/main/java/models/Book.java
@@ -1,4 +1,4 @@
-package model;
+package models;
 
 import java.util.Objects;
