+++ src/main/java/models/Car.java
@@ -1,4 +1,4 @@
-package model;
+package models;
 
 import java.util.Objects;
