+++ src/main/java/models/RootVegetable.java
@@ -1,4 +1,4 @@
-package model;
+package models;
 
 import java.util.Objects;
