--- src/main/java/strategy/impl/FillingListAuto.java
@@ -1,12 +1,0 @@
-package strategy.impl;
-
-import strategy.FillingList;
-
-import java.util.List;
-
-public class FillingListAuto implements FillingList {
-    @Override
-    public void fillList(List list) {
-
-    }
-}
